requirements.txt
boto>=2.49.0